backend/requirements.txt
fastapi
uvicorn
pydantic
pytest
pydantic[email]
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
PyJWT
xlsxwriter
pandas
openpyxl
python-multipart
